Huge refactor of script to rely on a src package with modules and subpackages
