+++ bin/pyquest_library_converter/src/enums.py
@@ -1,0 +1,7 @@
+import enum
+
+
+class OligoCasing(enum.Enum):
+    UPPER = "upper"
+    LOWER = "lower"
+    NONE = "unknown"
