+++ bin/pyquest_library_converter/src/exceptions.py
@@ -1,0 +1,10 @@
+class UserInterventionRequired(Exception):
+    pass
+
+
+class ValidationError(Exception):
+    pass
+
+
+class UndevelopedFeatureError(NotImplementedError):
+    pass
